Tweak openqa worker config
We can't have two tap worker hosts per instance as then openQA
assumes it can share related jobs between them, which we're not
set up for. Also the new boxes seem to be having some issues
with tap jobs. So let's make qa14 the only tap worker in prod
for now, and keep qa01 the tap worker for staging so we can
work out the problems.

Signed-off-by: Adam Williamson <awilliam@redhat.com>