openqa: make qa02 the tap worker host for production
I was holding off on this because of the theory that #1663318
was specific to the new worker hosts, but it turns out it was
not really, and we've fixed/worked around it now. qa01 has been
doing just fine in staging, so let's go with qa02 for prod; the
availability of more tap workers will help out.

Signed-off-by: Adam Williamson <awilliam@redhat.com>