bodhi: Disable gating.
Greenwave is returning HTTP 500 codes, and Bodhi currently
decides not to push updates when Greenwave returns a 500 code. I
plan to adjust Bodhi's logic such that it requires a 200 NOGO from
Greenwave to block an update, so that 500's allow the update to
proceed. However, that will take time and we have updates that are
getting blocked right now so this is a quick fix to unblock
packagers.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>