blockerbugs: changing SQLALCHEMY option in stg to reduce cron email
There was a change with a recent upgrade to flask-sqlalchemy so that
SQLALCHEMY_TRACK_MODIFICATIONS will no longer be enabled by default
in future releases. As we're not using it, I'm explicitly setting that
to False in the hopes that most/some of the cron emails go away if
the warning isn't emitted during blockerbugs sync.