[nagios/cloud-hosts] Fix virthost-aarch64
When adding a host to the cloud, it needs to be added to the inventory
file. The data in the file was organized by alphabet which meant that it
wasn't clear that hardware was also in this group. Reorganized the file.

Added data to the virthost data to make it clearer that it is like the
other cloud hosts. Removed it from the vpn mail group because it is not
on the vpn. We will need to create a cloud postfix group for these boxes.