proxy prerelease redirects: try a different way to disable/enable them
We had a messy setup that needed you to comment a block in one place
and uncomment it in the other. Instead lets try and use the template to
properly handle this and make a file without the redirect when we don't
want it enabled.

This only should affect staging now.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>