bodhi-backend: Add zchunk support to updates and updates-testing repositories
This re-adds zchunk support for the updates and updates-testing repositories
for both rpms and modularity.

Zchunk metadata was turned off due to a broken version of librepo that made it
out to stable, but a fixed version has been pushed and FESCo has decided[1] to
go ahead and turn this back on.

 1: https://pagure.io/fesco/issue/2116

Signed-off-by: Jonathan Dieter <jdieter@gmail.com>