Fedbages: put a higher threshold to fedmsg hub backlog nagios alerts.
The fedbages consumer has to consume many more messages and the query to datanomer
are slower. We need to allow the consumer to have a bigger backlog before raising an
alert.
Generally above 35000 message in the backlog it will be difficult for the consumer
to catch up, in that case it might be better to flush the backlog and restart from 0

Signed-off-by: Clement Verna <cverna@tutanota.com>