fas: backport stupid patch
Just backport and apply this patch during build.
The reasons are too insane for words: cherrypy-2.3.0 is actually not compatible with the Cookie library
from Python2.5 and up, because they add carriage returns.
It actually never broke in production because mod_wsgi daemon mode will handle and toss the \r's.
But gunicorn in EL6 has a backported security fix that throws the header entirely out.

This just took me way too deep down the stacks.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>