copr: don't install dnf-utils in provision builder
The dnf-utils package is not installable together with yum-utils
on Fedora 28, so even though it's needed for F30 - it would break
F28 spinup scripts.

To be precise - from F30 we should move to `copr-builder` package
anyways, which will bring all the necessary deps with itself; so
we'll remove the list entirely.