bodhi: Move callback into the general config.
I forgot that in TOML, the message headings capture all
settings below them until the next heading, and we want this
setting to be at the top level of the file.

Also, I thought ## would do a Jinja comment, but those comments
were getting rendered in the config, so I'm attempting a different
comment syntax.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>