bodhi: Install the composer in the consumer image.
The composer isn't actually used in OpenShift, but unfortunately
we forgot to make the import optional in Bodhi 4.0.0b1. This
results in a lot of unnecessary dependencies getting into the
container image, so I will make a future Bodhi release that makes
the composer import in the consumer optional. For now, let's just
make a bloated container.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>