oci-registry-prune: New playbook to prune old OCI images from the candidate registry.
This commit adds a new module and a new playbook used to delete images
from the candidate-registry that are older than 30 days.
The module looks for images tags that are older than 30 days and delete the
tags.
The playbook run the module and the registry garbage collection to reclaim
the disk space.

Signed-off-by: Clement Verna <cverna@tutanota.com>