openqa/dispatcher: first cut at supporting fedora-messaging
This is a big change to support fedora-messaging based consumers
instead of fedmsg ones in the openqa/dispatcher role. This needs
the fedora-messaging branch of fedora_openqa, until that is
merged to stable. For now this is all conditionalized so prod
will continue to use fedmsg. This also improves some variable
usage and documentation in the fedmsg path too.

Signed-off-by: Adam Williamson <awilliam@redhat.com>