openqa: partially revert @jcline's broker config change d76e5206
The way this was set up was intentional. For staging rabbitmq_
server is already set as the staging broker in openqa_stg group
vars; the setting here was an intentional override for it to use
the prod broker just for the scheduler queue. It has to do that
because it schedules jobs in response to composes and updates;
composes and updates almost never happen in the staging env, so
if we had the openQA staging scheduler listening to the staging
broker it'd almost never schedule any jobs. It has to listen
to the production broker and schedule jobs when real composes
and updates happen.

Still, perhaps the group var setting isn't working here, so
let's keep the explicit setting of the staging server here for
the other two queues and see if that helps that case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>