openqa playbook: rearrange, drop queue 'when' conditionals
This was kinda...a mess. I had just dumped the queue creation
in the staging NFS client play for some reason, and it still
had 'when' conditionals from when I only wanted the messaging
stuff to happen on staging. This cleans all that up and creates
separate plays for the queue creation and NFS client stuff.
The plays should also be in the right *order* now, so we get
the NFS mounts created and the AMQP queues set up *before* we
run the roles.