rabbit/queue/tasks: allow specifying custom write_queues
When creating a queue, the user is automatically created, but
there is no control over what queues the user can write to. A
user might need to write to queues other than ones prefixed with
their name (e.g. some users need to write to 'ci.*' queues).
This makes it possible to specify additional custom prefixes of
queues the user will be permitted to write to.

Signed-off-by: Adam Williamson <awilliam@redhat.com>