mirrorlist: make mirrorlist restart script look at .proto
In a previous commit I changed the non-templated version of
restart-mirrorlist-containers to look at the new protobuf based
mirrorlist cache file instead at the pkl file. It seems that nothing
uses the non-templated version of the script. This removes the unused
script and adapts the templated restart mirrorlist container script to
look at the protbuf based mirrorlist cache file.

Signed-off-by: Adrian Reber <adrian@lisas.de>