mirrorlist: fix last script referencing pkl
The script sync-old-pkl.yml (which is almost never used) was still
referencing pkl instead of the new proto extension for the mirrorlist
data exchange file. Now also using .proto. The name of the file is
unchanged which is not correct anymore. Not sure if the filename is
referenced in any documentation or SOP. Not changing it therefore right
now.

Signed-off-by: Adrian Reber <adrian@lisas.de>