openqa/dispatcher: restart all fm-consumer@ services daily
fedora-messaging has an annoying bug where consumers get stuck
allegedly working fine but actually not handling any messages:

https://github.com/fedora-infra/fedora-messaging/issues/208

to deal with this on our openQA servers, just have the dispatcher
role install a cron job which restarts all the services nightly.
Strictly speaking this shouldn't really be part of a specific
role like this, but this is the easiest way to do it and it's
not really that "wrong".

Signed-off-by: Adam Williamson <awilliam@redhat.com>